fix: budget_days float in VacationAccount schema + clock-status UTC elapsed time (#49, #50, #51)

+++ backend/app/schemas/reports.py
@@ -33,7 +33,7 @@ class VacationAccount(BaseModel):
     """Vacation account for a year."""
     year: int
     budget_hours: float
-    budget_days: int
+    budget_days: float
     used_hours: float
     used_days: float
     remaining_hours: float
